move request metadata functions inside rijksauth app
